+++ apps/docs/src/content/03-patterns/01-patterns/list-tile-table/examples/server-list.tsx
@@ -96,7 +96,7 @@ export default () => {
             )}
           </ServerList.Item>
           <ActionGroup>
-            <Button color="accent">
+            <Button color="success">
               <Text>Tarif bestellen</Text>
             </Button>
           </ActionGroup>