+++ apps/docs/src/content/04-components/actions/button/examples/hierarchy-solid-do.tsx
@@ -1,8 +1,0 @@
-import { Button } from "@mittwald/flow-react-components";
-
-<>
-  <Button color="secondary">
-    Speichern & weitere anlegen
-  </Button>
-  <Button color="accent">Anlegen</Button>
-</>;