+++ apps/docs/src/content/04-components/actions/button/examples/text-length-dont.tsx
@@ -1,7 +1,0 @@
-import { Button } from "@mittwald/flow-react-components";
-
-<Button color="accent">
-  Projekt und
-  <br />
-  Tarif anpassen
-</Button>;