+++ packages/components/src/components/ActionGroup/stories/Default.stories.tsx
@@ -5,6 +5,7 @@ import { Action } from "@/components/Action";
 import { sleep } from "@/lib/promises/sleep";
 import { Switch } from "@/components/Switch";
 import { Link } from "@/components/Link";
+import { Label } from "@/components/Label";
 import { useForm } from "react-hook-form";
 import { Form, SubmitButton } from "@/integrations/react-hook-form";
 
@@ -17,7 +18,7 @@ const meta: Meta<typeof ActionGroup> = {
   render: (props) => (
     <ActionGroup {...props}>
       <Action onAction={() => sleep(1500)}>
-        <Button color="accent">Recruit pilot</Button>
+        <Button color="success">Recruit pilot</Button>
       </Action>
       <Action>
         <Button slot="abort" variant="soft" color="secondary">
@@ -38,7 +39,7 @@ export const Size: Story = {
   render: (props) => (
     <ActionGroup {...props} size="s">
       <Link slot="abort">Forgot password?</Link>
-      <Button color="accent">Login</Button>
+      <Button color="success">Login</Button>
     </ActionGroup>
   ),
 };
@@ -54,7 +55,7 @@ export const PrimaryActionOnly: Story = {
 export const WithSecondaryAction: Story = {
   render: (props) => (
     <ActionGroup {...props}>
-      <Button slot="primary" color="accent">
+      <Button slot="primary" color="success">
         Add comlink address
       </Button>
       <Button slot="secondary" variant="soft" color="secondary">
@@ -70,7 +71,7 @@ export const WithSecondaryAction: Story = {
 export const WithExtraSecondaryAction: Story = {
   render: (props) => (
     <ActionGroup {...props}>
-      <Button slot="primary" color="accent">
+      <Button slot="primary" color="success">
         Add comlink address
       </Button>
       <Button slot="secondary" variant="soft" color="secondary">
@@ -92,7 +93,7 @@ export const PreserveOrder: Story = {
       <Button color="secondary" variant="soft">
         Abort
       </Button>
-      <Button color="accent">Create customer</Button>
+      <Button color="success">Create customer</Button>
       <Button color="secondary" variant="soft">
         Save as draft
       </Button>
@@ -104,7 +105,9 @@ export const WithSwitch: Story = {
   render: (props) => (
     <ActionGroup {...props}>
       <Button slot="secondary">Edit</Button>
-      <Switch slot="primary">Activate</Switch>
+      <Switch slot="primary">
+        <Label>Activate</Label>
+      </Switch>
     </ActionGroup>
   ),
 };
@@ -113,7 +116,7 @@ export const WithLink: Story = {
   render: (props) => (
     <ActionGroup {...props}>
       <Link slot="abort">Forgot password?</Link>
-      <Button color="accent">Login</Button>
+      <Button color="success">Login</Button>
     </ActionGroup>
   ),
 };