+++ packages/components/src/components/Chat/Chat.module.d.scss.ts
@@ -3,7 +3,7 @@ declare const classNames: {
   readonly messageThreadContainer: "messageThreadContainer";
   readonly fileCardList: "fileCardList";
   readonly controls: "controls";
-  readonly accentButton: "accentButton";
+  readonly successButton: "successButton";
   readonly button: "button";
 };
 export default classNames;